refactor: merge services into same library; moved files utils into phase2cli package
